+++ CHANGELOG.md
@@ -5,7 +5,8 @@
 - Rename SharedImage::from_rgb to from_image.
 - Add count method to ImageExt.
 - Modify visibility of internal image methods.
-- ImageExt::to_raw_rgb_data marked as unsafe.
+- ImageExt::to_rgb_data marked as unsafe.
+- Added ImageExt::to_rgb_image.
 - Update FLTK to commit b831848.
 
 ## [0.6.3] - 2020-06-13